// 10-ethers-again/react/src/models/Persons.ts
export interface PersonsInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "createPerson"
      | "getIndexList"
      | "numberOfPersons"
      | "personList"
      | "persons"
      | "removePerson"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "PersonCreated" | "PersonDeleted" | "PersonsIndex"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "createPerson",
    values: [string, BigNumberish, boolean]
  ): string;
  encodeFunctionData(
    functionFragment: "getIndexList",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "numberOfPersons",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "personList",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "persons",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "removePerson",
    values: [BigNumberish]
  ): string;

  decodeFunctionResult(
    functionFragment: "createPerson",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getIndexList",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "numberOfPersons",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "personList", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "persons", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "removePerson",
    data: BytesLike
  ): Result;
}

export namespace PersonCreatedEvent {
  export type InputTuple = [id: BigNumberish];
  export type OutputTuple = [id: bigint];
  export interface OutputObject {
    id: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace PersonDeletedEvent {
  export type InputTuple = [id: BigNumberish];
  export type OutputTuple = [id: bigint];
  export interface OutputObject {
    id: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace PersonsIndexEvent {
  export type InputTuple = [ids: BigNumberish[]];
  export type OutputTuple = [ids: bigint[]];
  export interface OutputObject {
    ids: bigint[];
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface Persons extends BaseContract {
  connect(runner?: ContractRunner | null): Persons;
  waitForDeployment(): Promise<this>;

  interface: PersonsInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  createPerson: TypedContractMethod<
    [_name: string, _age: BigNumberish, _isMarried: boolean],
    [void],
    "nonpayable"
  >;

  getIndexList: TypedContractMethod<[], [bigint[]], "view">;

  numberOfPersons: TypedContractMethod<[], [bigint], "view">;

  personList: TypedContractMethod<[arg0: BigNumberish], [bigint], "view">;

  persons: TypedContractMethod<
    [arg0: BigNumberish],
    [
      [bigint, string, bigint, boolean] & {
        id: bigint;
        name: string;
        age: bigint;
        isMarried: boolean;
      }
    ],
    "view"
  >;

  removePerson: TypedContractMethod<[id: BigNumberish], [void], "nonpayable">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "createPerson"
  ): TypedContractMethod<
    [_name: string, _age: BigNumberish, _isMarried: boolean],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "getIndexList"
  ): TypedContractMethod<[], [bigint[]], "view">;
  getFunction(
    nameOrSignature: "numberOfPersons"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "personList"
  ): TypedContractMethod<[arg0: BigNumberish], [bigint], "view">;
  getFunction(
    nameOrSignature: "persons"
  ): TypedContractMethod<
    [arg0: BigNumberish],
    [
      [bigint, string, bigint, boolean] & {
        id: bigint;
        name: string;
        age: bigint;
        isMarried: boolean;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "removePerson"
  ): TypedContractMethod<[id: BigNumberish], [void], "nonpayable">;

  getEvent(
    key: "PersonCreated"
  ): TypedContractEvent<
    PersonCreatedEvent.InputTuple,
    PersonCreatedEvent.OutputTuple,
    PersonCreatedEvent.OutputObject
  >;
  getEvent(
    key: "PersonDeleted"
  ): TypedContractEvent<
    PersonDeletedEvent.InputTuple,
    PersonDeletedEvent.OutputTuple,
    PersonDeletedEvent.OutputObject
  >;
  getEvent(
    key: "PersonsIndex"
  ): TypedContractEvent<
    PersonsIndexEvent.InputTuple,
    PersonsIndexEvent.OutputTuple,
    PersonsIndexEvent.OutputObject
  >;

  filters: {
    "PersonCreated(uint256)": TypedContractEvent<
      PersonCreatedEvent.InputTuple,
      PersonCreatedEvent.OutputTuple,
      PersonCreatedEvent.OutputObject
    >;
    PersonCreated: TypedContractEvent<
      PersonCreatedEvent.InputTuple,
      PersonCreatedEvent.OutputTuple,
      PersonCreatedEvent.OutputObject
    >;

    "PersonDeleted(uint256)": TypedContractEvent<
      PersonDeletedEvent.InputTuple,
      PersonDeletedEvent.OutputTuple,
      PersonDeletedEvent.OutputObject
    >;
    PersonDeleted: TypedContractEvent<
      PersonDeletedEvent.InputTuple,
      PersonDeletedEvent.OutputTuple,
      PersonDeletedEvent.OutputObject
    >;

    "PersonsIndex(uint256[])": TypedContractEvent<
      PersonsIndexEvent.InputTuple,
      PersonsIndexEvent.OutputTuple,
      PersonsIndexEvent.OutputObject
    >;
    PersonsIndex: TypedContractEvent<
      PersonsIndexEvent.InputTuple,
      PersonsIndexEvent.OutputTuple,
      PersonsIndexEvent.OutputObject
    >;
  };
}